[FIX] use python 2 map_async timeout hack to fix keyboardinterrupt
According the Python bug 8296 (https://bugs.python.org/issue8296), KeyboardInterrupt is not handled correctly for multiprocessing.Pool. If the interrupt happens within a child process, then the the interrupt kills the child but the remaining processes in the pool continue and execution hangs. The simplest way to fix this is to use a non-blocking map and timeout after longer than any code will ever run.